Solve using fromIntegral

+++ unit2/lesson11/consider-this.hs
@@ -1,1 +1,1 @@
-myAverage aList = sum aList / length aList
+myAverage aList = sum aList / fromIntegral (length aList)
